Renamed prop-types.js to propTypes.js.
